Extract modules to separate file.
